test: Move form methods from commands into helpers (#1322)
Closes #1309
