+++ projects/storefrontapp-e2e-cypress/cypress/support/commands.ts
@@ -1,3 +1,2 @@
 import 'cypress-testing-library/add-commands';
 import './general.commands';
-import './forms.commands';
